server/requirements.txt
flask
flask-jwt-extended
flask-cors
pymongo
bcrypt
marshmallow
python-dotenv
pytest
mongomock
